// apps/webapp/lib/types/treasure-types.ts
export interface BenchMarkDataItem {
  category: string;
  metricA: number;
  metricB: number;
  metricC: number;
}

export interface AllocationData {
  category: string;
  Harmonia: number;
  IndustryAverage: number;
}

export interface efficiencyData {
  name: string;
  score: number;
  color: string;
}